fix: service implementation for unix
