+++ internal/service/service_unix.go
@@ -1,0 +1,45 @@
+//go:build linux || darwin
+
+package service
+
+import (
+	"context"
+	"os"
+	"os/signal"
+	"syscall"
+)
+
+func Run(runner Runner) (int, error) {
+	// Create a channel to listen for termination signals
+	signalReceived := make(chan os.Signal, 1)
+	signal.Notify(signalReceived, os.Interrupt, syscall.SIGTERM)
+
+	// Make go routines for the channels
+	stop := make(chan struct{})
+	ctxStop, cancelStop := context.WithCancel(context.Background())
+	defer cancelStop()
+
+	go func() {
+		select {
+		case <-signalReceived:
+			stop <- struct{}{}
+		case <-ctxStop.Done():
+		}
+	}()
+
+	running := make(chan struct{})
+	ctxRunning, cancelRunning := context.WithCancel(context.Background())
+	defer cancelRunning()
+
+	go func() {
+		select {
+		case <-running:
+		case <-ctxRunning.Done():
+		}
+	}()
+
+	// Execute the runner
+	exitCode := runner.Execute(stop, running)
+
+	return int(exitCode), nil
+}
